Temporary tests sending JSON-info from client to server (successfull)
