+++ common/README.md
@@ -1,0 +1,15 @@
+
+
+### File structure
+
+src/: Source files, i.e. .ts (or .js if written in it).
+
+dist/: Compiled and transpiled code, i.e. .js.
+
+lib/: modules (node.js, express.js, etc.).
+
+tests/: test files.
+
+logs/: Generated log files.
+
+temp/: Temporary files.
