+++ server/src/server_axios.ts
@@ -1,4 +1,4 @@
-import express, { Request, Response } from 'express';
+/* import express, { Request, Response } from 'express';
 
 const app = express();
 const PORT: number = 3000;
@@ -10,7 +10,7 @@ app.get('/', (req: Request, res: Response) => {
   res.send('Hello, World! The server is working.');
   //res.send({test: "first test"});
 });
-/* 
+
 app.post('/send', (req: Request, res: Response) => {
   const { message } = req.body as { message?: unknown };
   if (!message || typeof message !== 'string') {
@@ -19,7 +19,7 @@ app.post('/send', (req: Request, res: Response) => {
   console.log('Received from client:', message);
   messages.push(message);
   res.json({ status: 'Message received', message });
-}); */
+});
 
 
 
@@ -32,4 +32,4 @@ if (require.main === module) {
   app.listen(PORT, () => {
     console.log(`Server is running on http://localhost:${PORT}`);
   });
-}
+} */
